Fix "." after converted license problem
In more verbose terms: Introduce lexicographically decreasing order
of license texts to have a simple greedy matching of the regular
expressions. Note that another solution would be to handle this in
the construction of the regular expressions; which might needed one
day if we are not only deadling with one license being the prefix of
another.