Sort TO-CLARIFY license statements
Or four statements have the same slightly problematic content: The say
to be LGPL-2.0-only OR LGPL-3.0-only or any later version that a proxy
decides. Yet for the proxy definition, they refer to the GPL version of
the license. The definitions of a proxy in both LGPL-3.0 and GPL-3.0 licenses
are equal exact that the first refers to the LGPL license and libraries and
the second to GPL and programs.
In my opinion the license text is clear enough that these statements mean
LGPL-2.0-only OR LGPL-3.0-only or LicenseRef-KDE-Accepted-LGPL